Storage & Project
- fixed release & game data persistance,
- fixed wrong usage of QList `resize`,
- handled unknown release case
